backend/pyproject.toml
[project]
name = "omniimage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.57.0",
    "llvmlite>=0.42.0",
    "numba>=0.59.0",
    "openai>=2.14.0",
    "peewee>=3.18.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pywebview>=6.1",
    "rembg>=2.0.72",
]
